rename HackMapProjection to MetricMapProjection
- and cleanups
