+++ src/org/osm2world/core/map_data/creation/MercatorProjection.java
@@ -21,11 +21,43 @@ final class MercatorProjection {
 	private static final double RATIO = R_MINOR / R_MAJOR;
 	private static final double ECCENT = sqrt(1.0 - (RATIO * RATIO));
 	private static final double COM = 0.5 * ECCENT;
+	public static final double EARTH_CIRCUMFERENCE = 40075016.686;
 
-	// public static double lonToX(double lon) {
-	// return toRadians(lon);
-	// }
+	/**
+	 *  Calculate earth circumference at given latitude.
+	 */
+	public static double earthCircumference(double latitude) {
+		return EARTH_CIRCUMFERENCE * Math.cos(toRadians(latitude));
+	}
 
+	/**
+	 * Convert longitude to Mercator projection (range [0..1]).
+	 */ 
+	public static double lonToX(double longitude) {
+		return (longitude + 180.0) / 360.0;
+	}
+
+	/**
+	 * Convert from Mercator projection (range [0..1]) to longitude.
+	 */
+	public static double xToLon(double x) {
+		return 360.0 * (x - 0.5);
+	}
+
+	/**
+	 * Convert latitude to Mercator projection (range [0..1]).
+	 */ 
+	public static double latToY(double latitude) {
+		double sinLat = Math.sin(toRadians(latitude));
+		return Math.log((1.0 + sinLat) / (1.0 - sinLat)) / (4.0 * Math.PI) + 0.5;
+	}
+
+	/**
+	 * Convert from Mercator projection (range [0..1]) to latitude.
+	 */
+	public static double yToLat(double y) {
+		return 360.0 * Math.atan(Math.exp((y - 0.5) * (2.0 * Math.PI))) / Math.PI - 90.0;
+	}
 	// This is for the Elliptical Mercator version
 	public static double latToYElliptical(double lat) {
 		lat = Math.min(89.5, Math.max(lat, -89.5));
@@ -36,14 +68,6 @@ final class MercatorProjection {
 		double ts = tan(0.5 * ((PI * 0.5) - phi)) / con;
 		return 0 - R_MAJOR * log(ts);
 	}
-
-	// public static double latToY(double lat) {
-	// return Math.log(Math.tan(PI / 4 + toRadians(lat) / 2));
-	// }
-	//
-	// public static double xToLon(double x) {
-	// return toDegrees(x);
-	// }
 
 	// This is for the Elliptical Mercator version
 	public static double yToLatElliptical(double y) {
@@ -59,38 +83,5 @@ final class MercatorProjection {
 			i++;
 		}
 		return toDegrees(phi);
-	}
-
-	// public static double yToLat(double y) {
-	// return toDegrees(2.0 * atan(exp(y)) - PI / 2);
-	// }
-
-	public static final double EARTH_CIRCUMFERENCE = 40075016.686;
-
-	/* calculate earth circumference at given latitude */
-	public static double earthCircumference(double latitude) {
-		return EARTH_CIRCUMFERENCE * Math.cos(toRadians(latitude));
-	}
-
-	/*
-	 * just replaced for testing as I know that these function reliably map
-	 * geocoords (within mercator limits)to 0..1 range - not having to search
-	 * for my logatrithm tables :)
-	 */
-	public static double lonToX(double lon) {
-		return (lon + 180.0) / 360.0;
-	}
-
-	public static double xToLon(double x) {
-		return 360.0 * (x - 0.5);
-	}
-
-	public static double latToY(double latitude) {
-		double sinLatitude = Math.sin(toRadians(latitude));
-		return Math.log((1 + sinLatitude) / (1 - sinLatitude)) / (4 * Math.PI) + 0.5;
-	}
-
-	public static double yToLat(double y) {
-		return 360 * Math.atan(Math.exp((y - 0.5) * (2 * Math.PI))) / Math.PI - 90;
 	}
 }
